Added utility functions module
